Squashed commit of the following:
commit 835d967f4618f92aab8650703f4bd938fb62d2bf
Author: bdj.dreamer <jkserena@naver.com>
Date:   Tue Oct 15 17:51:56 2024 +0900

    중복 코드 에러 수정

commit 2911c096736c6c4a43877061f0eae7753e6a1c21
Merge: 1aed46c 98e6b2b
Author: bdj.dreamer <jkserena@naver.com>
Date:   Tue Oct 15 17:48:25 2024 +0900

    Merge branch 'bdj'

commit 1aed46c71598b5fb422599d514c268e6541d9cb8
Author: bdj.dreamer <jkserena@naver.com>
Date:   Tue Oct 15 17:10:56 2024 +0900

    Update .factorypath

commit bd8177fa7dbbe2fd2787fb56d1cd3397339ea497
Merge: 28a96b3 321689d
Author: bdj.dreamer <jkserena@naver.com>
Date:   Tue Oct 15 17:10:11 2024 +0900

    Merge branch 'limwoochul'

commit 28a96b3f6aca6ca1cbf361d6d776d9b418d63c79
Merge: 2b7b14d 03819e0
Author: bdj.dreamer <jkserena@naver.com>
Date:   Tue Oct 15 17:06:00 2024 +0900

    Merge branch 'kbj'

commit 2b7b14d03210dda8a5a08eee24fe2738b4bee025
Merge: 8542042 4ab6a1a
Author: bdj.dreamer <jkserena@naver.com>
Date:   Tue Oct 15 17:04:17 2024 +0900

    Merge remote-tracking branch 'origin/psv_branch'

commit 321689d1fda5941a43fb1040f8160ffde8b01687
Author: limwoochul <wclim0319@naver.com>
Date:   Tue Oct 15 16:55:29 2024 +0900

    client 인터셉터 추가 + 마이페이지 접근 보안 강화

commit 8542042c56014335d11b8f345e165cf9760535c5
Merge: cc41330 6b855ea
Author: bdj.dreamer <jkserena@naver.com>
Date:   Tue Oct 15 16:49:37 2024 +0900

    Merge branch 'bdj'

commit cc41330d7e10eea7dcee027d504cc671dab6dc9d
Author: bdj.dreamer <jkserena@naver.com>
Date:   Tue Oct 15 16:49:25 2024 +0900

    dml 테스트 데이터 추가

commit 03819e038d99e8d6127a8a7908877f86eca6b107
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Tue Oct 15 16:00:42 2024 +0900

    직원 삭제 구현 완료

commit 322460f0c9bd45ac265d7f1bc0dd8670a529a9a1
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Tue Oct 15 15:31:07 2024 +0900

    테이블 api 적용 완료

commit 022c265013151b0f2ee0507886ed02c960522a60
Author: limwoochul <wclim0319@naver.com>
Date:   Tue Oct 15 15:20:20 2024 +0900

    admin 인터셉터 + 일부 메소드 보안강화(schedule 제외)

commit db960967e9b8c143763eef0a738e078e4fe64e20
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Tue Oct 15 14:00:33 2024 +0900

    테이블 api 적용

commit 8e1c51ab6cd8f5e648eaf8ec2dece105e01d7433
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Tue Oct 15 13:49:55 2024 +0900

    테이블 api 적용

commit 4ab6a1a694bb7dacd5c55355a4b9dcb2ecf393fb
Author: suvin5027 <123538283+suvin5027@users.noreply.github.com>
Date:   Tue Oct 15 12:28:32 2024 +0900

    [psv] 텍스트 변경

    [psv] 텍스트 변경

commit 42a65cee251d8c2b926cd20d8f35ed07fdfb2b2b
Author: suvin5027 <123538283+suvin5027@users.noreply.github.com>
Date:   Tue Oct 15 12:24:20 2024 +0900

    [psv] DDL, DML 추가

    payment_type의 pt_type을 pt_name으로 변경 후 pt_type을 새로 추가

commit 43e6216ce237e421e1b76dbbb103319e3137a2fe
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Tue Oct 15 12:18:37 2024 +0900

    코드 최적화 중

commit 112d0d27eb9e706aec2dfeb562933d57fc6471ff
Author: suvin5027 <123538283+suvin5027@users.noreply.github.com>
Date:   Tue Oct 15 12:04:41 2024 +0900

    [psv] 결제 기능 업데이트 중

    [psv]
    1. 결제 기능 업데이트 중
    2. header.jsp 들여쓰기 수정
    3. servlet-context.xml에서 로그인 해야 접근할 수 있는 경로에 paymentInsert 추가

commit 0ef652b62e1a74157a4543ece9e5b044d85be462
Author: limwoochul <wclim0319@naver.com>
Date:   Tue Oct 15 11:45:06 2024 +0900

    고객센터 일부 수정

    1. 리뷰등록, 리뷰수정에서 마이페이지 안보이던 부분 수정
    2. 마이페이지 문의내역 상세보기에서 답변 내용 추가
    3. 마이페이지 리뷰수정에 summernote 적용 안되어있는 부분 추가
    4. 마이페이지 리뷰등록에 취소버튼 추가

commit 8007f4905e9975d03927074e7dea2e6e145845a3
Author: limwoochul <wclim0319@naver.com>
Date:   Tue Oct 15 11:31:09 2024 +0900

    client폴더 jsp 폴더분류작업

commit a8ab563f8c8db909b31fdb40fd0d27f966fe4d61
Author: limwoochul <wclim0319@naver.com>
Date:   Tue Oct 15 10:59:40 2024 +0900

    admin폴더 jsp 폴더분류작업(schedule은 작업중이라 제외)

commit 02fc6507cbb08a8520b7d36760190bc1cf15c298
Author: limwoochul <wclim0319@naver.com>
Date:   Tue Oct 15 10:37:50 2024 +0900

    고객센터 1:1문의 빠진부분 구현 완료

commit 1b8d2f7948c0d925ee36759af90f52bf7bdcf754
Author: limwoochul <wclim0319@naver.com>
Date:   Tue Oct 15 10:37:31 2024 +0900

    지점 1:1문의내역 구현완료

commit e29ccc892871f5bfff7ec8b1323e4660b1c89dbc
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Tue Oct 15 10:14:43 2024 +0900

    테이블 api 필요한것만 적용

commit 7ce50366ec2bc9142b1d4b3aaae2318874186333
Author: suvin5027 <123538283+suvin5027@users.noreply.github.com>
Date:   Mon Oct 14 19:26:21 2024 +0900

    [psv] 결제 기능 구현 업데이트 중

    [psv] 결제 기능 구현 업데이트 중

commit e058d0e12416124f33ed97c810d9885108736b15
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Mon Oct 14 17:44:36 2024 +0900

    css 수정중

commit 217bcec9a1c2a19f939f6e7ee1ebc75d2cab166d
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Mon Oct 14 17:36:06 2024 +0900

    css 수정중

commit d03908f03cd3c5d8625d3445d3bdc0243b2c4f01
Author: suvin5027 <123538283+suvin5027@users.noreply.github.com>
Date:   Mon Oct 14 17:21:58 2024 +0900

    [psv] layout 수정

    [psv] layout 수정

commit ed5183576b1ae0728812292be6d784ae6671c229
Author: suvin5027 <123538283+suvin5027@users.noreply.github.com>
Date:   Mon Oct 14 17:21:17 2024 +0900

    [psv] 결제 기능 구현 업데이트 중

    [psv] 결제 기능 구현 업데이트 중

commit 17ee5a04d8ad92e78d9bcd1106306bea6864e4ae
Author: limwoochul <wclim0319@naver.com>
Date:   Mon Oct 14 16:08:42 2024 +0900

    지점 관리자 메뉴 페이지네이션 추가

commit 84d7d4875878d423de038b92e2083ade54d6fd74
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Mon Oct 14 14:15:56 2024 +0900

    코드 수정 중

commit 76976f9d06f5fdf1c8bf15e32cb200f19ab7e591
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Mon Oct 14 13:59:51 2024 +0900

    	 문의 내역 화면 이메일 전송 기능 추가 완료(config.properties 등록 해야함)

commit b0e3d0e22db74e1d405cf8c2b8feb462d34e95c8
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Mon Oct 14 13:55:33 2024 +0900

    문의 내역 화면 이메일 전송 기능 추가 완료(config.properties 등록 해야함)

commit 4b1bc9f43e23edeb15e9d33a149134c973b78141
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Mon Oct 14 13:14:22 2024 +0900

    문의 내역 화면 구현 완료

commit 2724d896aac926f8f60da093dba86ae6597f4b44
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Mon Oct 14 11:14:39 2024 +0900

    코드 최적화 중

commit bdf322055aab199fd0d73560bb4e306a6ea32b30
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Mon Oct 14 11:06:22 2024 +0900

    회원 관리 화면 구현 완료

commit b9cb1d5504bbaba0eefcf26b33feb8d718f6753c
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Mon Oct 14 09:56:18 2024 +0900

    코드 최적화 중

commit 0000c32cf918c4944c34866405ec4461bb90e978
Author: kbeomjun <98301971+kbeomjun@users.noreply.github.com>
Date:   Mon Oct 14 09:47:02 2024 +0900

    프로그램 관리 기능 구현 완료

commit 5253ab12d36adfa1edd4427a3d4620cd38164f2b
Author: suvin5027 <123538283+suvin5027@users.noreply.github.com>
Date:   Fri Oct 11 12:35:56 2024 +0900

    [psv] mapper 수정

    [psv] mapper 수정

commit 128f86c0c8afbc5fc3700d234a89a19f0f4bfaff
Author: suvin5027 <123538283+suvin5027@users.noreply.github.com>
Date:   Fri Oct 11 12:35:44 2024 +0900

    [psv] payment_category테이블 외래 키 제약 조건 추가

    [psv] payment_category테이블 외래 키 제약 조건 추가

+++ finalProject/src/main/java/kr/kh/fitness/controller/ClientController.java
@@ -17,6 +17,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import kr.kh.fitness.model.vo.BranchProgramScheduleVO;
 import kr.kh.fitness.model.vo.BranchVO;
@@ -24,6 +25,7 @@ import kr.kh.fitness.model.vo.InquiryTypeVO;
 import kr.kh.fitness.model.vo.MemberInquiryVO;
 import kr.kh.fitness.model.vo.MemberVO;
 import kr.kh.fitness.model.vo.PaymentVO;
+import kr.kh.fitness.model.vo.ProgramReservationVO;
 import kr.kh.fitness.model.vo.RefundVO;
 import kr.kh.fitness.model.vo.ReviewPostVO;
 import kr.kh.fitness.pagination.Criteria;
@@ -37,13 +39,7 @@ public class ClientController {
 	@Autowired
 	private ClientService clientService;
 	
-	@GetMapping("/mypage/{me_id}")
-	private String mypage(Model model, @PathVariable("me_id")String me_id) {
-		model.addAttribute("me_id", me_id);
-		
-		return "/client/mypage";
-	}
-	
+	//리뷰 게시판 목록
 	@GetMapping("/review/list")
 	public String reviewList(Model model, HttpSession session, Criteria cri) {
 		MemberVO user = (MemberVO)session.getAttribute("user");
@@ -56,9 +52,10 @@ public class ClientController {
 		model.addAttribute("reviewList", reviewList);
 		model.addAttribute("user", user);
 		model.addAttribute("pm", pm);
-		return "/client/reviewList";
+		return "/client/review/list";
 	}
 	
+	//리뷰 게시글 상세
 	@GetMapping("/review/detail/{rp_num}")
 	public String reviewDetail(Model model, @PathVariable("rp_num")int rp_num, HttpSession session, Criteria cri) {
 		
@@ -71,20 +68,25 @@ public class ClientController {
 		model.addAttribute("review", review);
 		model.addAttribute("user", user);
 		model.addAttribute("cri", cri);
-		return "/client/reviewDetail";
+		return "/client/review/detail";
 	}
 
-	@GetMapping("/review/insert/{me_id}")
-	public String reviewInsert(Model model, @PathVariable("me_id")String me_id) {
+	//리뷰 게시글 등록 get
+	@GetMapping("/review/insert")
+	public String reviewInsert(Model model, HttpSession session) {
 		//결제내역이 있는지 체크, 있으면 등록가능 없으면 등록불가
-		String msg = clientService.checkMemberPayment(me_id);
-		List<PaymentVO> paymentList = clientService.getPaymentListForReview(me_id);
+		MemberVO user = (MemberVO)session.getAttribute("user");
+		
+		String msg = clientService.checkMemberPayment(user.getMe_id());
+		
+		List<PaymentVO> paymentList = clientService.getPaymentListForReview(user.getMe_id());
 		List<BranchVO> branchList = clientService.getBranchList();
 		
 		if(msg == "") {
 			model.addAttribute("paymentList", paymentList);
 			model.addAttribute("branchList", branchList);
-			return "/client/reviewInsert";
+			model.addAttribute("user", user);
+			return "/client/review/insert";
 		} else {
 			model.addAttribute("url", "/client/review/list");
 			model.addAttribute("msg", msg);
@@ -92,6 +94,7 @@ public class ClientController {
 		}
 	}
 	
+	//리뷰 게시글 등록 post
 	@PostMapping("/review/insert")
 	public String reviewInsertPost(Model model, ReviewPostVO review, HttpSession session) {
 		
@@ -108,16 +111,26 @@ public class ClientController {
 		return "/main/message";
 	}
 	
+	//리뷰 게시글 수정 get
 	@GetMapping("/review/update/{rp_num}")
-	public String reviewUpdate(Model model, @PathVariable("rp_num")int rp_num) {
+	public String reviewUpdate(Model model, @PathVariable("rp_num")int rp_num, HttpSession session, RedirectAttributes redirectAttributes) {
 		ReviewPostVO review = clientService.getReviewPost(rp_num);
 		List<BranchVO> branchList = clientService.getBranchList();
+		MemberVO user = (MemberVO)session.getAttribute("user");
+		
+		if(!user.getMe_id().equals(review.getPa_me_id())) {
+	        redirectAttributes.addFlashAttribute("msg", "자신의 게시글만 접근 가능합니다.");
+	        return "redirect:/client/review/list";
+		}
+		
 		
 		model.addAttribute("review", review);
 		model.addAttribute("branchList", branchList);
-		return "/client/reviewUpdate";
+		model.addAttribute("user", user);
+		return "/client/review/update";
 	}
 	
+	//리뷰 게시글 수정 post
 	@PostMapping("/review/update")
 	public String reviewUpdatePost(Model model, ReviewPostVO review) {
 		
@@ -132,8 +145,16 @@ public class ClientController {
 		return "/main/message";
 	}
 	
+	//리뷰 게시글 삭제
 	@GetMapping("/review/delete/{rp_num}")
-	public String reviewDelete(Model model, @PathVariable("rp_num")int rp_num) {
+	public String reviewDelete(Model model, @PathVariable("rp_num")int rp_num, HttpSession session, RedirectAttributes redirectAttributes) {
+		ReviewPostVO review = clientService.getReviewPost(rp_num);
+		MemberVO user = (MemberVO)session.getAttribute("user");
+		
+		if(user == null || !user.getMe_id().equals(review.getPa_me_id())) {
+	        redirectAttributes.addFlashAttribute("msg", "자신의 게시글만 접근 가능합니다.");
+	        return "redirect:/client/review/list";
+		}		
 		
 		String msg = clientService.deleteReviewPost(rp_num);
 		
@@ -146,6 +167,7 @@ public class ClientController {
 		return "/main/message";
 	}
 	
+	//1:1문의 등록 get
 	@GetMapping("/inquiry/insert")
 	public String inquiryInsert(Model model, HttpSession session) {
 		MemberVO user = (MemberVO)session.getAttribute("user");
@@ -156,11 +178,36 @@ public class ClientController {
 		model.addAttribute("inquiryTypeList", inquiryTypeList);
 		model.addAttribute("branchList", branchList);
 		
-		return "/client/inquiryInsert";
+		return "/client/inquiry/insert";
 	}
 	
+	//1:1문의 등록 post
+	@PostMapping("/inquiry/insert")
+	public String inquiryInsertPost(Model model, MemberInquiryVO inquiry) {
+		
+		if(clientService.insertInquiry(inquiry)) {
+			model.addAttribute("msg", "등록에 성공했습니다.");
+		} else {
+			model.addAttribute("msg", "등록에 실패했습니다.");
+		}
+		model.addAttribute("url", "/client/inquiry/insert");
+		
+		return "/main/message";
+	}
+	
+	//마이페이지 스케줄 조회
 	@GetMapping("/mypage/schedule/{me_id}")
-	public String mypageSchedule(Model model, @PathVariable("me_id")String me_id, @RequestParam(value = "view", defaultValue = "present")String view, Criteria cri) {
+	public String mypageSchedule(Model model, @PathVariable("me_id")String me_id, 
+								@RequestParam(value = "view", defaultValue = "present")String view, 
+								Criteria cri, HttpSession session, RedirectAttributes redirectAttributes) {
+	    
+	    MemberVO user = (MemberVO) session.getAttribute("user");
+
+	    if (user == null || !user.getMe_id().equals(me_id)) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/schedule/" + user.getMe_id();
+	    }
+		
 		
 		cri.setPerPageNum(5);
 		
@@ -172,11 +219,21 @@ public class ClientController {
 		model.addAttribute("me_id", me_id);
 		model.addAttribute("pm", pm);
 		
-		return "/client/mypageSchedule";
+		return "/client/mypage/schedule";
 	}
 	
-	@GetMapping("/mypage/schedule/cancel/{pr_num}/{bs_num}/{me_id}")
-	public String mypageScheduleCancel(Model model, @PathVariable("pr_num")int pr_num, @PathVariable("bs_num")int bs_num, @PathVariable("me_id")String me_id) {
+	//마이페이지 스케줄 예약취소
+	@GetMapping("/mypage/schedule/cancel/{pr_num}/{bs_num}")
+	public String mypageScheduleCancel(Model model, @PathVariable("pr_num")int pr_num, @PathVariable("bs_num")int bs_num,
+										HttpSession session, RedirectAttributes redirectAttributes) {
+		
+	    MemberVO user = (MemberVO) session.getAttribute("user");
+	    ProgramReservationVO reservation = clientService.getReservation(pr_num);
+
+	    if (user == null || !user.getMe_id().equals(reservation.getPr_me_id())) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/schedule/" + user.getMe_id();
+	    }
 		
 		if(clientService.deleteReservation(pr_num)) {
 			clientService.updateScheduleCurrent(bs_num);
@@ -184,12 +241,20 @@ public class ClientController {
 		} else {
 			model.addAttribute("msg", "취소에 실패했습니다.");
 		}
-		model.addAttribute("url", "/client/mypage/schedule/" + me_id);
+		model.addAttribute("url", "/client/mypage/schedule/" + user.getMe_id());
 		return "/main/message";
 	}
 	
+	//마이페이지 회원권 내역
 	@GetMapping("/mypage/membership/{me_id}")
-	public String mypageMembership(Model model, @PathVariable("me_id")String me_id, Criteria cri) {
+	public String mypageMembership(Model model, @PathVariable("me_id")String me_id, Criteria cri, HttpSession session, RedirectAttributes redirectAttributes) {
+		
+	    MemberVO user = (MemberVO) session.getAttribute("user");
+
+	    if (user == null || !user.getMe_id().equals(me_id)) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/membership/" + user.getMe_id();
+	    }
 		
 		cri.setPerPageNum(5);
 		
@@ -200,21 +265,32 @@ public class ClientController {
 		model.addAttribute("paymentList", paymentList);
 		model.addAttribute("pm", pm);
 		
-		return "/client/mypageMembership";
+		return "/client/mypage/membership";
 	}
 	
-	@GetMapping("/mypage/review/insert/{pa_num}/{me_id}/{page}")
-	public String mypageReviewInsert(Model model, @PathVariable("pa_num")int pa_num, @PathVariable("me_id")String me_id, @PathVariable("page")int page) {
+	//마이페이지 회원권 -> 리뷰 작성 get
+	@GetMapping("/mypage/review/insert/{pa_num}/{page}")
+	public String mypageReviewInsert(Model model, @PathVariable("pa_num")int pa_num, @PathVariable("page")int page,
+										HttpSession session, RedirectAttributes redirectAttributes) {
+		
+	    MemberVO user = (MemberVO) session.getAttribute("user");
+	    PaymentVO payment = clientService.getpayment(pa_num);
+
+	    if (user == null || !user.getMe_id().equals(payment.getPa_me_id())) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/membership/" + user.getMe_id();
+	    }
 		
 		List<BranchVO> branchList = clientService.getBranchList();
 		
 		model.addAttribute("pa_num", pa_num);
-		model.addAttribute("me_id", me_id);
+		model.addAttribute("me_id", user.getMe_id());
 		model.addAttribute("page", page);
 		model.addAttribute("branchList", branchList);
-		return "/client/mypageReviewInsert";
+		return "/client/mypage/reviewInsert";
 	}
 	
+	//마이페이지 회원권 -> 리뷰 작성 post
 	@PostMapping("/mypage/review/insert")
 	public String mypageReviewInsertPost(Model model, ReviewPostVO review, String me_id, int page) {
 		
@@ -229,6 +305,7 @@ public class ClientController {
 		return "/main/message";
 	}
 	
+	//마이페이지 회원권 환불내역 상세조회
 	@ResponseBody
 	@GetMapping("/mypage/refundDetail")
 	public Map<String, Object> mypageRefundDetail(@RequestParam("pa_num") int pa_num){
@@ -241,8 +318,16 @@ public class ClientController {
 		return map;
 	}
 	
+	//마이페이지 리뷰게시글 내역 조회
 	@GetMapping("/mypage/review/list/{me_id}")
-	public String mypageReviewList(Model model, @PathVariable("me_id")String me_id, Criteria cri) {
+	public String mypageReviewList(Model model, @PathVariable("me_id")String me_id, Criteria cri, HttpSession session, RedirectAttributes redirectAttributes) {
+		
+	    MemberVO user = (MemberVO) session.getAttribute("user");
+
+	    if (user == null || !user.getMe_id().equals(me_id)) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/review/list/" + user.getMe_id();
+	    }
 		
 		cri.setPerPageNum(5);
 		cri.setType("id");
@@ -255,35 +340,50 @@ public class ClientController {
 		model.addAttribute("me_id", me_id);
 		model.addAttribute("pm", pm);
 		
-		return "/client/mypageReviewList";
+		return "/client/mypage/reviewList";
 	}
 	
-	@GetMapping("/mypage/review/detail/{rp_num}/{me_id}")
-	public String mypageReviewDetail(Model model, @PathVariable("rp_num")int rp_num, @PathVariable("me_id")String me_id, Criteria cri) {
+	//마이페이지 리뷰게시글 상세
+	@GetMapping("/mypage/review/detail/{rp_num}")
+	public String mypageReviewDetail(Model model, @PathVariable("rp_num")int rp_num, Criteria cri, HttpSession session, RedirectAttributes redirectAttributes) {
+		
+	    MemberVO user = (MemberVO) session.getAttribute("user");
+	    ReviewPostVO review = clientService.getReviewPost(rp_num);
+
+	    if (user == null || !user.getMe_id().equals(review.getPa_me_id())) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/review/list/" + user.getMe_id();
+	    }
 		
 		clientService.updateReviewPostView(rp_num);
 		
-		ReviewPostVO review = clientService.getReviewPost(rp_num);
-		
 		model.addAttribute("review", review);
-		model.addAttribute("me_id", me_id);
+		model.addAttribute("me_id", user.getMe_id());
 		model.addAttribute("cri", cri);
 		
-		return "/client/mypageReviewDetail";
+		return "/client/mypage/reviewDetail";
 	}
 	
-	@GetMapping("/mypage/review/update/{rp_num}/{me_id}")
-	public String mypageReviewUpdate(Model model, @PathVariable("rp_num")int rp_num, @PathVariable("me_id")String me_id) {
+	//마이페이지 리뷰게시글 수정 get
+	@GetMapping("/mypage/review/update/{rp_num}")
+	public String mypageReviewUpdate(Model model, @PathVariable("rp_num")int rp_num, HttpSession session, RedirectAttributes redirectAttributes) {
+		MemberVO user = (MemberVO) session.getAttribute("user");
 		ReviewPostVO review = clientService.getReviewPost(rp_num);
 		List<BranchVO> branchList = clientService.getBranchList();
 		
+	    if (user == null || !user.getMe_id().equals(review.getPa_me_id())) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/review/list/" + user.getMe_id();
+	    }
+		
 		model.addAttribute("review", review);
 		model.addAttribute("branchList", branchList);
-		model.addAttribute("me_id", me_id);
+		model.addAttribute("me_id", user.getMe_id());
 		
-		return "/client/mypageReviewUpdate";
+		return "/client/mypage/reviewUpdate";
 	}
 	
+	//마이페이지 리뷰게시글 수정 post
 	@PostMapping("/mypage/review/update")
 	public String mypageReviewUpdatePost(Model model, ReviewPostVO review, String me_id) {
 		
@@ -297,11 +397,17 @@ public class ClientController {
 		return "/main/message";
 	}
 	
+	//마이페이지 문의내역 목록
 	@GetMapping("/mypage/inquiry/list/{me_id}")
-	public String mypageInquiryList(Model model, @PathVariable("me_id")String me_id, Criteria cri) {
+	public String mypageInquiryList(Model model, @PathVariable("me_id")String me_id, Criteria cri, HttpSession session, RedirectAttributes redirectAttributes) {
 		
-		MemberVO user = clientService.getMember(me_id);
-		
+	    MemberVO user = (MemberVO) session.getAttribute("user");
+
+	    if (user == null || !user.getMe_id().equals(me_id)) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/inquiry/list/" + user.getMe_id();
+	    }
+	    
 		cri.setPerPageNum(3);
 		
 		List<MemberInquiryVO> inquiryList = clientService.getInquiryList(user.getMe_email(), cri);
@@ -311,27 +417,43 @@ public class ClientController {
 		model.addAttribute("inquiryList", inquiryList);
 		model.addAttribute("pm", pm);
 		
-		return "/client/mypageInquiryList";
+		return "/client/mypage/inquiryList";
 	}
 	
-	@GetMapping("/mypage/inquiry/detail/{mi_num}/{me_id}")
-	public String mypageInquiryDetail(Model model, @PathVariable("mi_num")int mi_num, @PathVariable("me_id")String me_id, int page) {
+	//마이페이지 문의내역 상세
+	@GetMapping("/mypage/inquiry/detail/{mi_num}")
+	public String mypageInquiryDetail(Model model, @PathVariable("mi_num")int mi_num, int page, HttpSession session, RedirectAttributes redirectAttributes) {
 		
+		MemberVO user = (MemberVO) session.getAttribute("user");
 		MemberInquiryVO inquiry = clientService.getInquiry(mi_num);
 		
+	    if (user == null || !user.getMe_email().equals(inquiry.getMi_email())) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/inquiry/list/" + user.getMe_id();
+	    }
+		
 		model.addAttribute("inquiry", inquiry);
-		model.addAttribute("me_id", me_id);
+		model.addAttribute("me_id", user.getMe_id());
 		model.addAttribute("page", page);
 		
-		return "/client/mypageInquiryDetail";
+		return "/client/mypage/inquiryDetail";
 	}
 	
+	//마이페이지 개인정보수정 시 비밀번호 확인 get
 	@GetMapping("/mypage/pwcheck/{me_id}")
-	public String mypagePwCheck(Model model, @PathVariable("me_id")String me_id) {
+	public String mypagePwCheck(Model model, @PathVariable("me_id")String me_id, HttpSession session, RedirectAttributes redirectAttributes) {
+	    MemberVO user = (MemberVO) session.getAttribute("user");
+
+	    if (user == null || !user.getMe_id().equals(me_id)) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/pwcheck/" + user.getMe_id();
+	    }
+	    
 		model.addAttribute("me_id", me_id);
-		return "/client/mypagePwCheck";
+		return "/client/mypage/pwCheck";
 	}
 	
+	//마이페이지 개인정보수정 시 비밀번호 확인 post
 	@PostMapping("/mypage/pwcheck")
 	public String mypagePwCheckPost(Model model, HttpSession session, MemberVO member) {
 		
@@ -347,8 +469,9 @@ public class ClientController {
 		}
 	}
 	
+	//마이페이지 개인정보수정 get
 	@GetMapping("/mypage/info/{me_id}")
-	public String mypageInfo(Model model, @PathVariable("me_id") String me_id, HttpSession session) {
+	public String mypageInfo(Model model, @PathVariable("me_id") String me_id, HttpSession session, RedirectAttributes redirectAttributes) {
 	    Boolean pwVerified = (Boolean) session.getAttribute("pwVerified");
 	    
 	    if (pwVerified == null || !pwVerified) {
@@ -356,13 +479,18 @@ public class ClientController {
 	        return "redirect:/client/mypage/pwcheck" + me_id;
 	    }
 	    
-	    MemberVO member = clientService.getMember(me_id);
+	    MemberVO user = (MemberVO) session.getAttribute("user");
+	    if (user == null || !user.getMe_id().equals(me_id)) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/pwcheck/" + user.getMe_id();
+	    }
 	    
-	    model.addAttribute("member", member);
+	    model.addAttribute("member", user);
 	    
-	    return "/client/mypageInfo";
+	    return "/client/mypage/info";
 	}
 	
+	//마이페이지 개인정보수정 email중복체크
 	@ResponseBody
     @PostMapping("/mypage/checkEmail")
     public boolean mypageCheckEmail(@RequestParam("email") String email) {
@@ -370,6 +498,7 @@ public class ClientController {
     	return clientService.isEmailDuplicate(email);
     }
 	
+	//마이페이지 개인정보수정 post
 	@PostMapping("/mypage/info/update")
 	public String mypageInfoUpdate(Model model, MemberVO member, HttpSession session, String birth) {
 	    
@@ -392,12 +521,22 @@ public class ClientController {
 		}
 	}
 	
+	//마이페이지 비밀번호 변경 get
 	@GetMapping("/mypage/pwchange/{me_id}")
-	public String mypagePwChange(Model model, @PathVariable("me_id")String me_id) {
+	public String mypagePwChange(Model model, @PathVariable("me_id")String me_id, HttpSession session, RedirectAttributes redirectAttributes) {
+	    
+		MemberVO user = (MemberVO) session.getAttribute("user");
+
+	    if (user == null || !user.getMe_id().equals(me_id)) {
+	        redirectAttributes.addFlashAttribute("msg", "잘못된 접근입니다.");
+	        return "redirect:/client/mypage/pwchange/" + user.getMe_id();
+	    }
+		
 		model.addAttribute("me_id", me_id);
-		return "/client/mypagePwChange";
+		return "/client/mypage/pwChange";
 	}
 	
+	//마이페이지 비밀번호 변경 post
 	@PostMapping("/mypage/pwchange/update")
 	public String updatePassword(Model model, @RequestParam("current_pw") String currentPw, @RequestParam("new_pw") String newPw, @RequestParam("me_id") String me_id, HttpSession session) {
